hermodr/requirements.txt
flask
flask_httpauth
paho.mqtt
influxdb